implemented further code splitting
