+++ src/utils.h
@@ -13,10 +13,7 @@ template <typename T> void print_vector(const T& v);
 
 template <typename T> void print_matrix(const T& m);
 
-std::tuple<vector<double>, vector<int>, bool, vector<double>> requestInputs ();
-
 vector<vector<double>> vec_2_mat(size_t n_x, size_t n_y, const vector<double>& x);
 
-void initiateProgram ();
 
 #endif
